fix: warning with different signedness

+++ launcher/modplatform/flame/FlameInstanceCreationTask.cpp
@@ -426,7 +426,7 @@ bool FlameCreationTask::createInstance()
         const uint64_t sysMiB = Sys::getSystemRam() / Sys::mebibyte;
         const uint64_t max = sysMiB * 0.9;
 
-        if (recommendedRAM > max) {
+        if (static_cast<uint64_t>(recommendedRAM) > max) {
             logWarning(tr("The recommended memory of the modpack exceeds 90% of your system RAM—reducing it from %1 MiB to %2 MiB!")
                            .arg(recommendedRAM)
                            .arg(max));
